+++ rentACar/src/main/java/com/etiya/rentACar/business/concretes/IndividualCustomerManager.java
@@ -12,8 +12,10 @@ import com.etiya.rentACar.business.dtos.IndividualCustomerSearchListDto;
 import com.etiya.rentACar.business.requests.individualCustomerRequests.CreateIndividualCustomerRequest;
 import com.etiya.rentACar.business.requests.individualCustomerRequests.DeleteIndividualCustomerRequest;
 import com.etiya.rentACar.business.requests.individualCustomerRequests.UpdateIndividualCustomerRequest;
+import com.etiya.rentACar.core.utilities.business.BusinessRules;
 import com.etiya.rentACar.core.utilities.mapping.ModelMapperService;
 import com.etiya.rentACar.core.utilities.results.DataResult;
+import com.etiya.rentACar.core.utilities.results.ErrorResult;
 import com.etiya.rentACar.core.utilities.results.Result;
 import com.etiya.rentACar.core.utilities.results.SuccessDataResult;
 import com.etiya.rentACar.core.utilities.results.SuccessResult;
@@ -53,6 +55,10 @@ public class IndividualCustomerManager implements IndividualCustomerService {
 
 	@Override
 	public Result update(UpdateIndividualCustomerRequest updateIndividualCustomerRequest) {
+		var result = BusinessRules.run(checkUserExists(updateIndividualCustomerRequest.getId()));
+		if (result!= null) {
+			return result;
+		}
 		IndividualCustomer individualCustomer = modelMapperService.forRequest().map(updateIndividualCustomerRequest, IndividualCustomer.class);
 		this.individualCustomerDao.save(individualCustomer);		
 		return new SuccessResult("Bireysel müşteri bilgileri güncellendi.");
@@ -60,8 +66,22 @@ public class IndividualCustomerManager implements IndividualCustomerService {
 
 	@Override
 	public Result delete(DeleteIndividualCustomerRequest deleteIndividualCustomerRequest) {
+		var result = BusinessRules.run(checkUserExists(deleteIndividualCustomerRequest.getId()));
+		if (result!= null) {
+			return result;
+		}
 		this.individualCustomerDao.delete(this.individualCustomerDao.getById(deleteIndividualCustomerRequest.getId()));
 		return new SuccessResult("Bireysel müşteri bilgileri silindi.");
+	}
+	
+	private Result checkUserExists(int id) {
+		var result = this.individualCustomerDao.existsById(id);
+		if (!result) {
+			return new ErrorResult("Kullanıcı bulunamadı.");
+			
+		}
+		
+		return new SuccessResult("Kullanıcı bulundu.");
 	}
 
 }
